backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
pyotp==2.9.0
qrcode[pil]==7.4.2
aiohttp==3.9.1
psutil==5.9.6
requests==2.31.0
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2